refactor(rewrite): move mod.rs test module to separate tests.rs file
Extract ~500-line inline test module into crates/rskim/src/cmd/rewrite/tests.rs
and replace with `#[cfg(test)] mod tests;`. Promote has_compound_operators,
collect_input_tokens, and SegmentClassification to pub(super) so the external
test file can reach them via `use super::*`.
